Fix compile time warning. No functional change
git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@511887 13f79535-47bb-0310-9956-ffa450edef68

+++ jk/native/common/jk_map.c
@@ -365,7 +365,7 @@ int jk_map_read_property(jk_map_t *m, const char *str, int allow_duplicates, jk_
             trim(v);
             if (strlen(v) && strlen(prp)) {
                 const char *oldv;
-                int off = strlen(prp) - JK_MAP_REFERENCE_SZ;
+                int off = (int)strlen(prp) - (int)JK_MAP_REFERENCE_SZ;
                 /* check the worker properties */
                 if (off <= 0 || strncmp(&prp[off], JK_MAP_REFERENCE, JK_MAP_REFERENCE_SZ) ) {
                     if (!jk_is_valid_property(prp)) {
